@@ -2,19 +2,35 @@
 
 import postDAO from '../DAO/postDAO';
 
-function create (context){
-    async function query(){
-        let result = await postDAO.query();
-        if(result){
+function create(context) {
+    async function createNewOrUpdate(data) {
+        let result = await postDAO.createNewOrUpdate(data);
+        if (result) {
             return result;
         }
     }
-    
+
+    async function query() {
+        let result = await postDAO.query();
+        if (result) {
+            return result;
+        }
+    }
+
+    async function get(id) {
+        let result = await postDAO.get(id);
+        if (result) {
+            return result;
+        }
+    }
+
     return {
-        query: query
+        createNewOrUpdate: createNewOrUpdate,
+        query: query,
+        get: get,
     };
 }
 
-export default{
+export default {
     create: create
 };
