@@ -1,5 +1,159 @@
 (window["webpackJsonp"] = window["webpackJsonp"] || []).push([["main"],{
 
+/***/ "./node_modules/raw-loader/index.js!./src/app/app.component.html":
+/*!**************************************************************!*\
+  !*** ./node_modules/raw-loader!./src/app/app.component.html ***!
+  \**************************************************************/
+/*! no static exports found */
+/***/ (function(module, exports) {
+
+module.exports = "<!--The content below is only a placeholder and can be replaced.-->\r\n<navbar></navbar>\r\n<router-outlet></router-outlet>\r\n"
+
+/***/ }),
+
+/***/ "./node_modules/raw-loader/index.js!./src/app/components/add-post/add-post.component.html":
+/*!***************************************************************************************!*\
+  !*** ./node_modules/raw-loader!./src/app/components/add-post/add-post.component.html ***!
+  \***************************************************************************************/
+/*! no static exports found */
+/***/ (function(module, exports) {
+
+module.exports = "<div class=\"col-lg-9 col-md-9 col-sm-9\">\n\t<h4>Dodaj post:</h4>\n\t<form (ngSubmit)=\"submit()\">\n    \t<div class=\"form-group\">\n        \t<label class=\"col-form-label\">Tytuł:</label>\n        \t<input name=\"title\" type=\"text\" class=\"form-control\" placeholder=\"Wprowadź tytuł\" [(ngModel)]=\"post.title\">\n    \t</div>\n\n    \t<div class=\"form-group\">\n        \t<label class=\"col-form-label\">Adres obrazu:</label>\n        \t<input name=\"url\" type=\"text\" class=\"form-control\" placeholder=\"URL\" [(ngModel)]=\"post.url\">\n    \t</div>\n\n    \t<div class=\"form-group\">\n        \t<label class=\"col-form-label\">Treść:</label>\n        \t<textarea name=\"content\" type=\"text\" class=\"form-control\" placeholder=\"Treść\" [(ngModel)]=\"post.content\" rows=\"10\"></textarea>\n    \t</div>\n\n    \t<div class=\"form-group button\">\n        \t<button type=\"submit\" class=\"btn btn-info\">Dodaj</button>\n    \t</div>\n\t</form>\n</div>\n"
+
+/***/ }),
+
+/***/ "./node_modules/raw-loader/index.js!./src/app/components/blog-home/blog-home.component.html":
+/*!*****************************************************************************************!*\
+  !*** ./node_modules/raw-loader!./src/app/components/blog-home/blog-home.component.html ***!
+  \*****************************************************************************************/
+/*! no static exports found */
+/***/ (function(module, exports) {
+
+module.exports = "<search-bar (name)=\"getName($event)\">\r\n</search-bar>\r\n<blog [filterText]=\"filterText\">\r\n</blog>\r\n"
+
+/***/ }),
+
+/***/ "./node_modules/raw-loader/index.js!./src/app/components/blog-item-details/blog-item-details.component.html":
+/*!*********************************************************************************************************!*\
+  !*** ./node_modules/raw-loader!./src/app/components/blog-item-details/blog-item-details.component.html ***!
+  \*********************************************************************************************************/
+/*! no static exports found */
+/***/ (function(module, exports) {
+
+module.exports = "<img [src]=\"image\" >\r\n\t<p>\r\n {{ text}}\r\n</p>\r\n"
+
+/***/ }),
+
+/***/ "./node_modules/raw-loader/index.js!./src/app/components/blog-item-image/blog-item-image.component.html":
+/*!*****************************************************************************************************!*\
+  !*** ./node_modules/raw-loader!./src/app/components/blog-item-image/blog-item-image.component.html ***!
+  \*****************************************************************************************************/
+/*! no static exports found */
+/***/ (function(module, exports) {
+
+module.exports = "<div class=\"card-img-top\">\r\n<img [src]=\"image\">\r\n</div>"
+
+/***/ }),
+
+/***/ "./node_modules/raw-loader/index.js!./src/app/components/blog-item-text/blog-item-text.component.html":
+/*!***************************************************************************************************!*\
+  !*** ./node_modules/raw-loader!./src/app/components/blog-item-text/blog-item-text.component.html ***!
+  \***************************************************************************************************/
+/*! no static exports found */
+/***/ (function(module, exports) {
+
+module.exports = "<p class=\"card text-center\">\r\n  {{ text | summary:35 }}\r\n</p>\r\n\r\n<a [routerLink]=\"['/blog/detail/',id]\" class=\"btn btn-primary\">\r\n  Więcej\r\n </a>\r\n "
+
+/***/ }),
+
+/***/ "./node_modules/raw-loader/index.js!./src/app/components/blog-item/blog-item.component.html":
+/*!*****************************************************************************************!*\
+  !*** ./node_modules/raw-loader!./src/app/components/blog-item/blog-item.component.html ***!
+  \*****************************************************************************************/
+/*! no static exports found */
+/***/ (function(module, exports) {
+
+module.exports = "<blog-item-image [image]=\"image\"></blog-item-image>\r\n<blog-item-text [id]=\"id\" [text]=\"text\"></blog-item-text>"
+
+/***/ }),
+
+/***/ "./node_modules/raw-loader/index.js!./src/app/components/blog/blog.component.html":
+/*!*******************************************************************************!*\
+  !*** ./node_modules/raw-loader!./src/app/components/blog/blog.component.html ***!
+  \*******************************************************************************/
+/*! no static exports found */
+/***/ (function(module, exports) {
+
+module.exports = "<selectize></selectize>\r\n<div *ngFor=\"let item of items$ | filter: filterText\" class=\"card-deck\">\r\n  <blog-item [id]=\"item.id\" [image]=\"item.image\" [text]=\"item.title\">\r\n    \r\n  </blog-item>\r\n</div>"
+
+/***/ }),
+
+/***/ "./node_modules/raw-loader/index.js!./src/app/components/contact/contact.component.html":
+/*!*************************************************************************************!*\
+  !*** ./node_modules/raw-loader!./src/app/components/contact/contact.component.html ***!
+  \*************************************************************************************/
+/*! no static exports found */
+/***/ (function(module, exports) {
+
+module.exports = "  <div class=\"navbar\" id=\"navbar\"></div>\r\n\r\n  <form>\r\n      <div class=\"container\">\r\n          <div class=\"form-group\">\r\n              <label for=\"firstName\">First name</label>\r\n              <input type=\"text\" class=\"form-control\" id=\"exampleInputEmail1\" aria-describedby=\"emailHelp\"\r\n                  placeholder=\"Enter first name\">\r\n          </div>\r\n\r\n          <div class=\"form-group\">\r\n              <label for=\"firstName\">Last name</label>\r\n              <input type=\"text\" class=\"form-control\" id=\"exampleInputEmail1\" aria-describedby=\"emailHelp\"\r\n                  placeholder=\"Enter last name\">\r\n          </div>\r\n\r\n          <div class=\"form-group\">\r\n              <label for=\"exampleInputEmail1\">Email address</label>\r\n              <input type=\"email\" class=\"form-control\" id=\"exampleInputEmail1\" aria-describedby=\"emailHelp\"\r\n                  placeholder=\"Enter email\">\r\n              <small id=\"emailHelp\" class=\"form-text text-muted\">We'll never share your email with anyone\r\n                  else.</small>\r\n          </div>\r\n\r\n\r\n          <div class=\"form-check form-check-inline\">\r\n              <input class=\"form-check-input\" type=\"radio\" name=\"exampleRadios\" id=\"man\" value=\"option1\" checked>\r\n              <label class=\"form-check-label\" for=\"man\">\r\n                  Man\r\n              </label>\r\n          </div>\r\n\r\n          <div class=\"form-check form-check-inline\">\r\n              <input class=\"form-check-input\" type=\"radio\" name=\"exampleRadios\" id=\"woman\" value=\"option2\">\r\n              <label class=\"form-check-label\" for=\"woman\">\r\n                  Woman\r\n              </label>\r\n          </div>\r\n\r\n          <div class=\"form-group\">\r\n              <label for=\"exampleFormControlTextarea1\">Opis</label>\r\n              <textarea class=\"form-control\" id=\"exampleFormControlTextarea1\" rows=\"3\"></textarea>\r\n          </div>\r\n\r\n          <div class=\"form-group\">\r\n              <label for=\"customRange1\">Age</label>\r\n              <select class=\"custom-select\">\r\n                  <option selected>choose age</option>\r\n                  <option value=\"1\">>18</option>\r\n                  <option value=\"2\">18-25</option>\r\n                  <option value=\"3\">25-50</option>\r\n                  <option value=\"4\"><50</option> \r\n              </select> \r\n          </div> \r\n\r\n          <div class=\"form-group\">\r\n              <label for=\"inputState\">Marital status</label>\r\n              <select id=\"inputState\" class=\"form-control\">\r\n                  <option selected>single</option>\r\n                  <option>married</option>\r\n                  <option>divorced</option>\r\n                  <option>widow</option>\r\n              </select>\r\n          </div>\r\n\r\n          <div>\r\n              <button type=\"submit\" class=\"btn btn-primary\">Submit</button>\r\n          \r\n              <button type=\"reset\" class=\"btn btn-primary\">Reset</button>\r\n          </div>\r\n      </div>\r\n\r\n      <div class=\"scroll\">\r\n          <i class=\"fa fa-arrow-circle-up\" aria-hidden=\"true\">\r\n          <button class=\"btn btn-success\">Do góry</button>\r\n          </i>\r\n      </div>\r\n       \r\n  </form>\r\n"
+
+/***/ }),
+
+/***/ "./node_modules/raw-loader/index.js!./src/app/components/home/home.component.html":
+/*!*******************************************************************************!*\
+  !*** ./node_modules/raw-loader!./src/app/components/home/home.component.html ***!
+  \*******************************************************************************/
+/*! no static exports found */
+/***/ (function(module, exports) {
+
+module.exports = "  <div class=\"navbar\" id=\"navbar\"></div>\r\n    \r\n    <div class=\"row\">\r\n        <div class=\"col-9\">Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Hendrerit dolor magna eget est lorem ipsum. Duis at consectetur lorem donec massa. Feugiat in fermentum posuere urna nec. Elementum tempus egestas sed sed risus pretium quam vulputate dignissim. Orci porta non pulvinar neque laoreet suspendisse interdum consectetur. Ligula ullamcorper malesuada proin libero nunc. Non sodales neque sodales ut etiam sit. Scelerisque eleifend donec pretium vulputate sapien nec. Felis donec et odio pellentesque diam volutpat commodo sed. Enim blandit volutpat maecenas volutpat blandit aliquam etiam erat velit.\r\n\r\n                Consequat mauris nunc congue nisi vitae suscipit tellus. Natoque penatibus et magnis dis parturient. Bibendum ut tristique et egestas quis ipsum suspendisse ultrices. Maecenas volutpat blandit aliquam etiam. Varius vel pharetra vel turpis nunc eget lorem dolor sed. Volutpat est velit egestas dui id ornare arcu odio ut. Ut sem viverra aliquet eget sit amet tellus cras. Eget felis eget nunc lobortis. Faucibus in ornare quam viverra orci sagittis eu. Aliquam purus sit amet luctus venenatis lectus magna fringilla. Netus et malesuada fames ac turpis egestas. Non arcu risus quis varius quam quisque.\r\n                \r\n                Vitae proin sagittis nisl rhoncus mattis. Eget est lorem ipsum dolor sit amet consectetur adipiscing elit. Lacus sed turpis tincidunt id aliquet risus. Penatibus et magnis dis parturient montes nascetur ridiculus mus mauris. Malesuada fames ac turpis egestas integer eget. Semper quis lectus nulla at volutpat diam ut. Purus non enim praesent elementum facilisis. Porttitor lacus luctus accumsan tortor posuere ac. Accumsan in nisl nisi scelerisque eu. Nibh tellus molestie nunc non. Ut morbi tincidunt augue interdum velit euismod. Tellus integer feugiat scelerisque varius morbi enim nunc. Id aliquet lectus proin nibh nisl condimentum. Lacinia quis vel eros donec ac. Purus in massa tempor nec feugiat. Tellus rutrum tellus pellentesque eu. Pretium viverra suspendisse potenti nullam ac tortor vitae purus. Donec ultrices tincidunt arcu non sodales neque sodales ut. Nisl tincidunt eget nullam non. Sit amet consectetur adipiscing elit ut aliquam purus sit.\r\n                \r\n                Nisi scelerisque eu ultrices vitae auctor eu augue ut. Odio eu feugiat pretium nibh. Amet consectetur adipiscing elit ut aliquam purus. Ullamcorper malesuada proin libero nunc consequat interdum varius sit. Egestas congue quisque egestas diam in arcu cursus. Ut diam quam nulla porttitor massa id neque aliquam vestibulum. Scelerisque eu ultrices vitae auctor eu augue ut lectus. Sed enim ut sem viverra aliquet eget sit amet tellus. Sit amet consectetur adipiscing elit pellentesque habitant morbi tristique senectus. Massa enim nec dui nunc mattis. Mattis aliquam faucibus purus in massa tempor nec feugiat nisl. Malesuada fames ac turpis egestas maecenas. Tortor posuere ac ut consequat semper viverra. Praesent tristique magna sit amet purus gravida. In fermentum et sollicitudin ac orci. Pretium fusce id velit ut tortor pretium viverra suspendisse potenti. Quis lectus nulla at volutpat diam ut venenatis tellus. Tempor orci eu lobortis elementum nibh tellus molestie. Nunc eget lorem dolor sed viverra.\r\n                \r\n                Facilisi cras fermentum odio eu feugiat. Ornare arcu dui vivamus arcu felis bibendum. Curabitur gravida arcu ac tortor dignissim convallis aenean. Venenatis tellus in metus vulputate eu. Eros donec ac odio tempor orci dapibus ultrices in. Nunc sed blandit libero volutpat sed cras. Mauris rhoncus aenean vel elit scelerisque mauris pellentesque pulvinar. Id diam vel quam elementum pulvinar. At lectus urna duis convallis convallis tellus id interdum velit. Rutrum tellus pellentesque eu tincidunt tortor aliquam nulla facilisi. Adipiscing elit duis tristique sollicitudin nibh sit amet commodo. Pharetra pharetra massa massa ultricies mi quis. Imperdiet proin fermentum leo vel. Justo nec ultrices dui sapien eget mi proin. Convallis a cras semper auctor neque vitae tempus. Morbi tristique senectus et netus et. Nulla facilisi nullam vehicula ipsum a arcu. Nisl nisi scelerisque eu ultrices vitae auctor eu.\r\n                \r\n                Curabitur gravida arcu ac tortor. Pulvinar proin gravida hendrerit lectus. Ut faucibus pulvinar elementum integer enim neque volutpat ac. Ornare arcu dui vivamus arcu felis bibendum ut. Leo vel fringilla est ullamcorper eget nulla facilisi. Lectus magna fringilla urna porttitor rhoncus dolor. Venenatis lectus magna fringilla urna porttitor rhoncus. Eu sem integer vitae justo eget. Vulputate sapien nec sagittis aliquam malesuada bibendum. Arcu ac tortor dignissim convallis aenean. Morbi tincidunt augue interdum velit euismod. Condimentum id venenatis a condimentum vitae sapien pellentesque. Ultrices eros in cursus turpis massa tincidunt dui ut ornare. Ac turpis egestas sed tempus urna et pharetra pharetra massa. Sit amet consectetur adipiscing elit ut. Ultrices neque ornare aenean euismod. Eget lorem dolor sed viverra ipsum nunc aliquet.\r\n                \r\n                Scelerisque eu ultrices vitae auctor eu augue ut lectus. Non quam lacus suspendisse faucibus. Diam sollicitudin tempor id eu nisl nunc. Nulla aliquet porttitor lacus luctus. Vulputate mi sit amet mauris. Egestas integer eget aliquet nibh. Vestibulum lectus mauris ultrices eros in cursus turpis massa. Nibh tellus molestie nunc non blandit massa enim nec. Et molestie ac feugiat sed lectus vestibulum mattis ullamcorper. Id cursus metus aliquam eleifend mi in nulla posuere. Integer eget aliquet nibh praesent.\r\n                \r\n                Pellentesque eu tincidunt tortor aliquam nulla. Sed sed risus pretium quam vulputate dignissim suspendisse in est. Sem fringilla ut morbi tincidunt augue interdum velit. Ultrices in iaculis nunc sed augue lacus viverra vitae congue. Netus et malesuada fames ac turpis egestas. Hendrerit gravida rutrum quisque non tellus orci ac. Amet luctus venenatis lectus magna fringilla urna. Vitae sapien pellentesque habitant morbi tristique senectus et. Nullam non nisi est sit amet facilisis. Feugiat nibh sed pulvinar proin. Id cursus metus aliquam eleifend mi in. Pulvinar pellentesque habitant morbi tristique. Fringilla est ullamcorper eget nulla facilisi etiam dignissim diam quis. Cursus eget nunc scelerisque viverra. Eu volutpat odio facilisis mauris sit. Dapibus ultrices in iaculis nunc. Neque ornare aenean euismod elementum. Justo eget magna fermentum iaculis eu. Quam pellentesque nec nam aliquam sem et tortor. Magna fringilla urna porttitor rhoncus.\r\n                \r\n                Pulvinar mattis nunc sed blandit libero volutpat. Laoreet sit amet cursus sit amet dictum sit. Pharetra magna ac placerat vestibulum lectus mauris ultrices eros. Placerat duis ultricies lacus sed turpis. Volutpat maecenas volutpat blandit aliquam. Aliquam faucibus purus in massa tempor nec feugiat nisl pretium. Morbi tristique senectus et netus et malesuada fames ac turpis. Risus nec feugiat in fermentum posuere. Nec tincidunt praesent semper feugiat nibh sed pulvinar proin. Nibh nisl condimentum id venenatis. Pharetra sit amet aliquam id diam. Elementum curabitur vitae nunc sed velit dignissim. Diam maecenas ultricies mi eget. Tortor vitae purus faucibus ornare suspendisse sed nisi lacus sed. Egestas quis ipsum suspendisse ultrices gravida dictum fusce. Mauris ultrices eros in cursus turpis massa tincidunt. Accumsan tortor posuere ac ut. Fames ac turpis egestas integer eget aliquet nibh praesent tristique.\r\n                \r\n                Elementum nisi quis eleifend quam adipiscing vitae. Pulvinar neque laoreet suspendisse interdum. Cursus vitae congue mauris rhoncus. Molestie a iaculis at erat pellentesque. Quis enim lobortis scelerisque fermentum dui faucibus in ornare quam. Nec nam aliquam sem et tortor. Ullamcorper eget nulla facilisi etiam dignissim diam quis enim lobortis. Dui nunc mattis enim ut. Leo a diam sollicitudin tempor. Enim nunc faucibus a pellentesque sit amet porttitor eget dolor. Senectus et netus et malesuada fames ac turpis egestas maecenas. At lectus urna duis convallis convallis tellus id. Neque vitae tempus quam pellentesque nec. Nulla porttitor massa id neque aliquam vestibulum morbi. Mauris nunc congue nisi vitae suscipit tellus mauris a diam. Mauris cursus mattis molestie a iaculis.\r\n                \r\n                Justo eget magna fermentum iaculis eu non. Vulputate mi sit amet mauris. Vulputate dignissim suspendisse in est ante in nibh. Justo donec enim diam vulputate ut pharetra sit amet. Elementum nisi quis eleifend quam. Ipsum faucibus vitae aliquet nec. Sit amet consectetur adipiscing elit duis tristique. Id consectetur purus ut faucibus pulvinar elementum integer enim. Nunc sed velit dignissim sodales ut eu. Cras pulvinar mattis nunc sed blandit libero volutpat sed cras. Lacus luctus accumsan tortor posuere ac ut consequat semper viverra. Pellentesque habitant morbi tristique senectus. Amet nulla facilisi morbi tempus. Porta nibh venenatis cras sed. Lectus quam id leo in vitae turpis. Vivamus arcu felis bibendum ut tristique et egestas quis. Accumsan lacus vel facilisis volutpat est velit. Erat nam at lectus urna duis convallis convallis.\r\n                \r\n                Consequat mauris nunc congue nisi vitae suscipit tellus. In ante metus dictum at tempor commodo ullamcorper. Eget arcu dictum varius duis at consectetur. At urna condimentum mattis pellentesque id nibh tortor. Purus gravida quis blandit turpis cursus in. Dui vivamus arcu felis bibendum ut tristique et egestas. Nulla facilisi cras fermentum odio eu feugiat. Magna ac placerat vestibulum lectus mauris ultrices. Tellus pellentesque eu tincidunt tortor aliquam nulla facilisi cras fermentum. Iaculis urna id volutpat lacus laoreet non curabitur gravida. Aliquet risus feugiat in ante metus dictum at tempor. Diam volutpat commodo sed egestas. Eget lorem dolor sed viverra ipsum nunc aliquet bibendum enim. Nulla facilisi etiam dignissim diam quis enim. Arcu vitae elementum curabitur vitae. Hendrerit gravida rutrum quisque non tellus orci ac auctor. Ornare arcu odio ut sem nulla pharetra. Sollicitudin aliquam ultrices sagittis orci a scelerisque purus semper eget. Et leo duis ut diam quam nulla.\r\n                \r\n                Non consectetur a erat nam at lectus urna duis. Gravida neque convallis a cras semper auctor neque vitae. Sed velit dignissim sodales ut eu sem integer vitae justo. Facilisis leo vel fringilla est ullamcorper. Semper quis lectus nulla at volutpat diam ut venenatis. Scelerisque eleifend donec pretium vulputate sapien nec. Mauris pharetra et ultrices neque ornare aenean euismod elementum. Est placerat in egestas erat imperdiet sed euismod nisi. Nunc vel risus commodo viverra maecenas accumsan lacus vel facilisis. Amet consectetur adipiscing elit duis tristique sollicitudin nibh. Maecenas sed enim ut sem. Sagittis eu volutpat odio facilisis mauris.\r\n                \r\n                Egestas congue quisque egestas diam in arcu cursus. Sit amet purus gravida quis blandit turpis cursus in. Egestas egestas fringilla phasellus faucibus scelerisque eleifend donec pretium vulputate. Lobortis scelerisque fermentum dui faucibus in ornare quam. Turpis egestas pretium aenean pharetra magna. Amet facilisis magna etiam tempor orci eu lobortis elementum nibh. Amet est placerat in egestas erat imperdiet sed euismod. Sed vulputate mi sit amet mauris commodo. Sagittis aliquam malesuada bibendum arcu vitae elementum. Nunc sed blandit libero volutpat sed. Fringilla est ullamcorper eget nulla facilisi etiam. Mi tempus imperdiet nulla malesuada pellentesque. Lobortis elementum nibh tellus molestie nunc. Feugiat nisl pretium fusce id velit ut. Vel pharetra vel turpis nunc eget lorem. Arcu non sodales neque sodales ut etiam.\r\n                \r\n                Mauris augue neque gravida in fermentum et. Porttitor rhoncus dolor purus non enim. Consequat mauris nunc congue nisi vitae suscipit. Eget gravida cum sociis natoque. Phasellus vestibulum lorem sed risus ultricies tristique nulla. Velit dignissim sodales ut eu sem integer vitae justo. Aliquet sagittis id consectetur purus ut faucibus. Donec massa sapien faucibus et molestie. Quam lacus suspendisse faucibus interdum posuere lorem. Semper auctor neque vitae tempus quam pellentesque nec nam aliquam. Fermentum dui faucibus in ornare quam viverra orci sagittis. A lacus vestibulum sed arcu non odio euismod.\r\n                \r\n                Aliquam ut porttitor leo a diam. Leo integer malesuada nunc vel risus commodo. Tempor orci eu lobortis elementum nibh. At risus viverra adipiscing at in tellus integer feugiat scelerisque. Nibh cras pulvinar mattis nunc sed blandit libero volutpat. Viverra accumsan in nisl nisi scelerisque eu ultrices vitae auctor. Adipiscing enim eu turpis egestas pretium aenean pharetra magna ac. In dictum non consectetur a erat nam at. Integer malesuada nunc vel risus commodo viverra maecenas. Platea dictumst quisque sagittis purus sit. Aliquet enim tortor at auctor urna nunc id cursus. Scelerisque varius morbi enim nunc faucibus a. Sapien nec sagittis aliquam malesuada bibendum arcu vitae elementum curabitur. Urna condimentum mattis pellentesque id nibh tortor. Justo laoreet sit amet cursus sit amet dictum. Ridiculus mus mauris vitae ultricies leo integer. Ac ut consequat semper viverra nam libero. Vel facilisis volutpat est velit egestas dui. Vehicula ipsum a arcu cursus. Nunc congue nisi vitae suscipit tellus mauris a.\r\n                \r\n                Consequat id porta nibh venenatis cras sed felis. Eu volutpat odio facilisis mauris sit amet. A erat nam at lectus urna duis convallis convallis tellus. Pretium viverra suspendisse potenti nullam. Sed ullamcorper morbi tincidunt ornare massa eget egestas purus. Ut tristique et egestas quis. Neque viverra justo nec ultrices dui sapien eget mi. Aenean et tortor at risus. Et leo duis ut diam quam nulla. Suspendisse faucibus interdum posuere lorem ipsum dolor sit amet. Enim lobortis scelerisque fermentum dui faucibus in ornare quam. Parturient montes nascetur ridiculus mus mauris vitae ultricies leo. Sagittis aliquam malesuada bibendum arcu vitae. A diam maecenas sed enim. Aenean et tortor at risus viverra. Pulvinar pellentesque habitant morbi tristique senectus et netus et.\r\n                \r\n                Felis eget velit aliquet sagittis id. Sed vulputate mi sit amet mauris commodo quis imperdiet massa. Dui ut ornare lectus sit amet est placerat in. Convallis a cras semper auctor neque vitae tempus quam. Risus at ultrices mi tempus. Egestas pretium aenean pharetra magna. Massa enim nec dui nunc. In nibh mauris cursus mattis molestie a iaculis at erat. Elementum nisi quis eleifend quam adipiscing. Eget arcu dictum varius duis. Scelerisque varius morbi enim nunc faucibus a pellentesque sit. Lorem sed risus ultricies tristique nulla aliquet enim. Velit ut tortor pretium viverra. Vitae semper quis lectus nulla at volutpat diam ut. Erat pellentesque adipiscing commodo elit at imperdiet dui accumsan. Eget duis at tellus at urna.\r\n                \r\n                Accumsan tortor posuere ac ut consequat. Lacus vestibulum sed arcu non odio euismod lacinia at. Diam quam nulla porttitor massa id neque aliquam. Ac tortor vitae purus faucibus ornare suspendisse sed nisi lacus. Sed turpis tincidunt id aliquet risus feugiat. Mattis molestie a iaculis at erat pellentesque adipiscing commodo. Non sodales neque sodales ut etiam sit amet. Amet mattis vulputate enim nulla aliquet porttitor. Tristique risus nec feugiat in fermentum posuere urna. Elit scelerisque mauris pellentesque pulvinar pellentesque habitant morbi. Sit amet porttitor eget dolor morbi non.\r\n                \r\n                Enim ut tellus elementum sagittis vitae et leo duis. Eget lorem dolor sed viverra ipsum. Orci ac auctor augue mauris augue neque gravida in fermentum. Metus vulputate eu scelerisque felis imperdiet proin fermentum leo. Sed vulputate mi sit amet mauris. Vitae congue eu consequat ac felis donec. Odio ut sem nulla pharetra diam sit amet. Eget nulla facilisi etiam dignissim diam quis. Lorem mollis aliquam ut porttitor leo a diam sollicitudin tempor. In egestas erat imperdiet sed euismod nisi porta lorem mollis. Est lorem ipsum dolor sit amet consectetur adipiscing elit. Vitae congue mauris rhoncus aenean vel elit scelerisque. In pellentesque massa placerat duis ultricies lacus sed turpis tincidunt. Consequat id porta nibh venenatis cras. Amet risus nullam eget felis eget nunc lobortis. Ipsum faucibus vitae aliquet nec. Enim facilisis gravida neque convallis a. Dictumst quisque sagittis purus sit amet volutpat consequat mauris nunc.</div>\r\n        <div class=\"col-3 s-bg\">\r\n            <h3>Szybki kontakt</h3>\r\n            \r\n            <a href=\"#\"><i class=\"fab fa-facebook-square\"></i></a>\r\n            <a href=\"#\"><i class=\"fab fa-instagram\"></i></a>\r\n            <a href=\"#\"><i class=\"fas fa-at\"></i></a>\r\n        </div>\r\n    </div>\r\n\r\n    <div class=\"scroll\">\r\n            <button class=\"btn btn-success\" id=\"myButton\">\r\n                <i class=\"fa fa-arrow-circle-up\" aria-hidden=\"true\"></i>\r\n                 Do góry\r\n            </button>\r\n    </div>\r\n             \r\n    <div class=\"footer\">Copyrights © Sebastian Pulak | 2019</div>\r\n"
+
+/***/ }),
+
+/***/ "./node_modules/raw-loader/index.js!./src/app/components/quiz/quiz.component.html":
+/*!*******************************************************************************!*\
+  !*** ./node_modules/raw-loader!./src/app/components/quiz/quiz.component.html ***!
+  \*******************************************************************************/
+/*! no static exports found */
+/***/ (function(module, exports) {
+
+module.exports = "\r\n  <div class=\"navbar\" id=\"navbar\"></div>\r\n<div class=\"row\"></div>\r\n<div class=\"container list\">\r\n\r\n    Question: <span class=\"index\"></span>1<h4 class=\"question\"></h4>\r\n\r\n    Answers:\r\n    <ul class=\"list-group\">\r\n        <li class=\"list-group-item\"></li>\r\n        <li class=\"list-group-item\"></li>\r\n        <li class=\"list-group-item\"></li>\r\n        <li class=\"list-group-item\"></li>\r\n    </ul>\r\n\r\n    <div class=\"buttons\">\r\n        <button class=\"btn btn-danger previous\">Previous</button>\r\n        <button class=\"btn btn-success next\">Next</button>\r\n    </div>\r\n\r\n    <h4>Your score: <span class=\"score\">0</span>/20 points.</h4>\r\n</div>\r\n<div class=\"container results\" style=\"display: none\">\r\n    <div class=\"btn-group-custom\">\r\n        <button class=\"btn btn-success btn-custom restart\">Retake quiz</button>\r\n    </div>\r\n    <div class=\"score\">\r\n        <table class=\"table table-striped\">\r\n            <thead>\r\n            <tr>\r\n                <th>#</th>\r\n                <th>Point/20</th>\r\n                <th>Average</th>\r\n            </tr>\r\n            </thead>\r\n            <tbody>\r\n            <tr>\r\n                <th scope=\"row\">Your score</th>\r\n                <td class=\"userScorePoint\"></td>\r\n                <td class=\"average\"></td>\r\n            </tr>\r\n            </tbody>\r\n        </table>\r\n    </div>\r\n</div>\r\n\r\n\r\n"
+
+/***/ }),
+
+/***/ "./node_modules/raw-loader/index.js!./src/app/components/selectize/selectize.component.html":
+/*!*****************************************************************************************!*\
+  !*** ./node_modules/raw-loader!./src/app/components/selectize/selectize.component.html ***!
+  \*****************************************************************************************/
+/*! no static exports found */
+/***/ (function(module, exports) {
+
+module.exports = "\r\n<input #input type=\"text\">\r\n\r\n<ul>\r\n  <li *ngFor=\"let p of posts$\">\r\n\t<a href=\"#\" [routerLink]=\"['/blog/detail/', p.id]\"> {{ p.text }}</a></li>\r\n</ul>\r\n"
+
+/***/ }),
+
+/***/ "./node_modules/raw-loader/index.js!./src/app/navbar/navbar.component.html":
+/*!************************************************************************!*\
+  !*** ./node_modules/raw-loader!./src/app/navbar/navbar.component.html ***!
+  \************************************************************************/
+/*! no static exports found */
+/***/ (function(module, exports) {
+
+module.exports = "<nav>\r\n    <ul class=\"nav\">\r\n        <li class=\"nav-item\"><a class=\"nav-link active\" [routerLink]=\"['/']\">Home</a></li>\r\n        <li class=\"nav-item\"><a class=\"nav-link\" [routerLink]=\"['/blog']\">Blog</a></li>\r\n        <li class=\"nav-item\"><a class=\"nav-link\" [routerLink]=\"['/quiz']\">Quiz</a></li>\r\n        <li class=\"nav-item\"><a class=\"nav-link\" [routerLink]=\"['/contact']\">Contact</a></li>\r\n        <li class=\"nav-item\"><a class=\"nav-link\" [routerLink]=\"['/add-post']\">Add post</a></li>\r\n        <li class=\"nav-item\"><a class=\"nav-link\" [routerLink]=\"['/blog/detail/:id']\">Blog item details</a></li>\r\n    </ul>\r\n</nav>\r\n\r\n"
+
+/***/ }),
+
+/***/ "./node_modules/raw-loader/index.js!./src/app/shared/search-bar/search-bar.component.html":
+/*!***************************************************************************************!*\
+  !*** ./node_modules/raw-loader!./src/app/shared/search-bar/search-bar.component.html ***!
+  \***************************************************************************************/
+/*! no static exports found */
+/***/ (function(module, exports) {
+
+module.exports = "<input [(ngModel)]=\"filterText\"\r\n(ngModelChange)=\"sendFilter($event)\"\r\n      \tplaceholder=\"Wpisz nazwę...\" id=\"search\" type=\"search\"\r\nclass=\"form-control\">\r\n"
+
+/***/ }),
+
 /***/ "./src/$$_lazy_route_resource lazy recursive":
 /*!**********************************************************!*\
   !*** ./src/$$_lazy_route_resource lazy namespace object ***!
@@ -41,6 +195,8 @@ __webpack_require__.r(__webpack_exports__);
 /* harmony import */ var _components_quiz_quiz_component__WEBPACK_IMPORTED_MODULE_5__ = __webpack_require__(/*! ./components/quiz/quiz.component */ "./src/app/components/quiz/quiz.component.ts");
 /* harmony import */ var _components_blog_item_details_blog_item_details_component__WEBPACK_IMPORTED_MODULE_6__ = __webpack_require__(/*! ./components/blog-item-details/blog-item-details.component */ "./src/app/components/blog-item-details/blog-item-details.component.ts");
 /* harmony import */ var _components_blog_home_blog_home_component__WEBPACK_IMPORTED_MODULE_7__ = __webpack_require__(/*! ./components/blog-home/blog-home.component */ "./src/app/components/blog-home/blog-home.component.ts");
+/* harmony import */ var _components_add_post_add_post_component__WEBPACK_IMPORTED_MODULE_8__ = __webpack_require__(/*! ./components/add-post/add-post.component */ "./src/app/components/add-post/add-post.component.ts");
+
 
 
 
@@ -69,6 +225,10 @@ var routes = [
     {
         path: 'blog/detail/:id',
         component: _components_blog_item_details_blog_item_details_component__WEBPACK_IMPORTED_MODULE_6__["BlogItemDetailsComponent"]
+    },
+    {
+        path: 'add-post',
+        component: _components_add_post_add_post_component__WEBPACK_IMPORTED_MODULE_8__["AddPostComponent"]
     }
 ];
 var AppRoutingModule = /** @class */ (function () {
@@ -98,17 +258,6 @@ module.exports = "\n/*# sourceMappingURL=data:application/json;base64,eyJ2ZXJzaW
 
 /***/ }),
 
-/***/ "./src/app/app.component.html":
-/*!************************************!*\
-  !*** ./src/app/app.component.html ***!
-  \************************************/
-/*! no static exports found */
-/***/ (function(module, exports) {
-
-module.exports = "<!--The content below is only a placeholder and can be replaced.-->\r\n<navbar></navbar>\r\n<router-outlet></router-outlet>\r\n"
-
-/***/ }),
-
 /***/ "./src/app/app.component.ts":
 /*!**********************************!*\
   !*** ./src/app/app.component.ts ***!
@@ -130,7 +279,7 @@ var AppComponent = /** @class */ (function () {
     AppComponent = tslib__WEBPACK_IMPORTED_MODULE_0__["__decorate"]([
         Object(_angular_core__WEBPACK_IMPORTED_MODULE_1__["Component"])({
             selector: 'app-root',
-            template: __webpack_require__(/*! ./app.component.html */ "./src/app/app.component.html"),
+            template: __webpack_require__(/*! raw-loader!./app.component.html */ "./node_modules/raw-loader/index.js!./src/app/app.component.html"),
             styles: [__webpack_require__(/*! ./app.component.css */ "./src/app/app.component.css")]
         })
     ], AppComponent);
@@ -173,6 +322,8 @@ __webpack_require__.r(__webpack_exports__);
 /* harmony import */ var _components_blog_home_blog_home_component__WEBPACK_IMPORTED_MODULE_19__ = __webpack_require__(/*! ./components/blog-home/blog-home.component */ "./src/app/components/blog-home/blog-home.component.ts");
 /* harmony import */ var _pipes_filter_pipe__WEBPACK_IMPORTED_MODULE_20__ = __webpack_require__(/*! ./pipes/filter.pipe */ "./src/app/pipes/filter.pipe.ts");
 /* harmony import */ var _components_selectize_selectize_component__WEBPACK_IMPORTED_MODULE_21__ = __webpack_require__(/*! ./components/selectize/selectize.component */ "./src/app/components/selectize/selectize.component.ts");
+/* harmony import */ var _components_add_post_add_post_component__WEBPACK_IMPORTED_MODULE_22__ = __webpack_require__(/*! ./components/add-post/add-post.component */ "./src/app/components/add-post/add-post.component.ts");
+
 
 
 
@@ -215,7 +366,8 @@ var AppModule = /** @class */ (function () {
                 _shared_search_bar_search_bar_component__WEBPACK_IMPORTED_MODULE_17__["SearchBarComponent"],
                 _components_blog_home_blog_home_component__WEBPACK_IMPORTED_MODULE_19__["BlogHomeComponent"],
                 _pipes_filter_pipe__WEBPACK_IMPORTED_MODULE_20__["FilterPipe"],
-                _components_selectize_selectize_component__WEBPACK_IMPORTED_MODULE_21__["SelectizeComponent"]
+                _components_selectize_selectize_component__WEBPACK_IMPORTED_MODULE_21__["SelectizeComponent"],
+                _components_add_post_add_post_component__WEBPACK_IMPORTED_MODULE_22__["AddPostComponent"]
             ],
             imports: [
                 _angular_platform_browser__WEBPACK_IMPORTED_MODULE_1__["BrowserModule"],
@@ -236,6 +388,68 @@ var AppModule = /** @class */ (function () {
 
 /***/ }),
 
+/***/ "./src/app/components/add-post/add-post.component.css":
+/*!************************************************************!*\
+  !*** ./src/app/components/add-post/add-post.component.css ***!
+  \************************************************************/
+/*! no static exports found */
+/***/ (function(module, exports) {
+
+module.exports = "\n/*# sourceMappingURL=data:application/json;base64,eyJ2ZXJzaW9uIjozLCJzb3VyY2VzIjpbXSwibmFtZXMiOltdLCJtYXBwaW5ncyI6IiIsImZpbGUiOiJzcmMvYXBwL2NvbXBvbmVudHMvYWRkLXBvc3QvYWRkLXBvc3QuY29tcG9uZW50LmNzcyJ9 */"
+
+/***/ }),
+
+/***/ "./src/app/components/add-post/add-post.component.ts":
+/*!***********************************************************!*\
+  !*** ./src/app/components/add-post/add-post.component.ts ***!
+  \***********************************************************/
+/*! exports provided: AddPostComponent */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+__webpack_require__.r(__webpack_exports__);
+/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, "AddPostComponent", function() { return AddPostComponent; });
+/* harmony import */ var tslib__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! tslib */ "./node_modules/tslib/tslib.es6.js");
+/* harmony import */ var _angular_core__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! @angular/core */ "./node_modules/@angular/core/fesm5/core.js");
+/* harmony import */ var _angular_router__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! @angular/router */ "./node_modules/@angular/router/fesm5/router.js");
+/* harmony import */ var src_app_services_data_service__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! src/app/services/data.service */ "./src/app/services/data.service.ts");
+
+
+
+
+var AddPostComponent = /** @class */ (function () {
+    function AddPostComponent(dataService, router) {
+        this.dataService = dataService;
+        this.router = router;
+        this.post = {
+            title: '',
+            url: '',
+            content: '',
+        };
+    }
+    AddPostComponent.prototype.ngOnInit = function () {
+    };
+    AddPostComponent.prototype.submit = function () {
+        var _this = this;
+        this.dataService.createOrUpdate(this.post).subscribe(function (f) {
+            _this.router.navigate(['/blog']);
+        });
+    };
+    AddPostComponent = tslib__WEBPACK_IMPORTED_MODULE_0__["__decorate"]([
+        Object(_angular_core__WEBPACK_IMPORTED_MODULE_1__["Component"])({
+            selector: 'add-post',
+            template: __webpack_require__(/*! raw-loader!./add-post.component.html */ "./node_modules/raw-loader/index.js!./src/app/components/add-post/add-post.component.html"),
+            styles: [__webpack_require__(/*! ./add-post.component.css */ "./src/app/components/add-post/add-post.component.css")]
+        }),
+        tslib__WEBPACK_IMPORTED_MODULE_0__["__metadata"]("design:paramtypes", [src_app_services_data_service__WEBPACK_IMPORTED_MODULE_3__["DataService"], _angular_router__WEBPACK_IMPORTED_MODULE_2__["Router"]])
+    ], AddPostComponent);
+    return AddPostComponent;
+}());
+
+
+
+/***/ }),
+
 /***/ "./src/app/components/blog-home/blog-home.component.css":
 /*!**************************************************************!*\
   !*** ./src/app/components/blog-home/blog-home.component.css ***!
@@ -244,17 +458,6 @@ var AppModule = /** @class */ (function () {
 /***/ (function(module, exports) {
 
 module.exports = "\n/*# sourceMappingURL=data:application/json;base64,eyJ2ZXJzaW9uIjozLCJzb3VyY2VzIjpbXSwibmFtZXMiOltdLCJtYXBwaW5ncyI6IiIsImZpbGUiOiJzcmMvYXBwL2NvbXBvbmVudHMvYmxvZy1ob21lL2Jsb2ctaG9tZS5jb21wb25lbnQuY3NzIn0= */"
-
-/***/ }),
-
-/***/ "./src/app/components/blog-home/blog-home.component.html":
-/*!***************************************************************!*\
-  !*** ./src/app/components/blog-home/blog-home.component.html ***!
-  \***************************************************************/
-/*! no static exports found */
-/***/ (function(module, exports) {
-
-module.exports = "<search-bar (name)=\"getName($event)\">\r\n</search-bar>\r\n<blog [filterText]=\"filterText\">\r\n</blog>\r\n"
 
 /***/ }),
 
@@ -283,7 +486,7 @@ var BlogHomeComponent = /** @class */ (function () {
     BlogHomeComponent = tslib__WEBPACK_IMPORTED_MODULE_0__["__decorate"]([
         Object(_angular_core__WEBPACK_IMPORTED_MODULE_1__["Component"])({
             selector: 'blog-home',
-            template: __webpack_require__(/*! ./blog-home.component.html */ "./src/app/components/blog-home/blog-home.component.html"),
+            template: __webpack_require__(/*! raw-loader!./blog-home.component.html */ "./node_modules/raw-loader/index.js!./src/app/components/blog-home/blog-home.component.html"),
             styles: [__webpack_require__(/*! ./blog-home.component.css */ "./src/app/components/blog-home/blog-home.component.css")]
         }),
         tslib__WEBPACK_IMPORTED_MODULE_0__["__metadata"]("design:paramtypes", [])
@@ -303,17 +506,6 @@ var BlogHomeComponent = /** @class */ (function () {
 /***/ (function(module, exports) {
 
 module.exports = "\n/*# sourceMappingURL=data:application/json;base64,eyJ2ZXJzaW9uIjozLCJzb3VyY2VzIjpbXSwibmFtZXMiOltdLCJtYXBwaW5ncyI6IiIsImZpbGUiOiJzcmMvYXBwL2NvbXBvbmVudHMvYmxvZy1pdGVtLWRldGFpbHMvYmxvZy1pdGVtLWRldGFpbHMuY29tcG9uZW50LmNzcyJ9 */"
-
-/***/ }),
-
-/***/ "./src/app/components/blog-item-details/blog-item-details.component.html":
-/*!*******************************************************************************!*\
-  !*** ./src/app/components/blog-item-details/blog-item-details.component.html ***!
-  \*******************************************************************************/
-/*! no static exports found */
-/***/ (function(module, exports) {
-
-module.exports = "<img [src]=\"image\" >\r\n\t<p>\r\n {{ text}}\r\n</p>\r\n"
 
 /***/ }),
 
@@ -361,7 +553,7 @@ var BlogItemDetailsComponent = /** @class */ (function () {
     BlogItemDetailsComponent = tslib__WEBPACK_IMPORTED_MODULE_0__["__decorate"]([
         Object(_angular_core__WEBPACK_IMPORTED_MODULE_1__["Component"])({
             selector: 'blog-item-details',
-            template: __webpack_require__(/*! ./blog-item-details.component.html */ "./src/app/components/blog-item-details/blog-item-details.component.html"),
+            template: __webpack_require__(/*! raw-loader!./blog-item-details.component.html */ "./node_modules/raw-loader/index.js!./src/app/components/blog-item-details/blog-item-details.component.html"),
             styles: [__webpack_require__(/*! ./blog-item-details.component.css */ "./src/app/components/blog-item-details/blog-item-details.component.css")]
         }),
         tslib__WEBPACK_IMPORTED_MODULE_0__["__metadata"]("design:paramtypes", [src_app_services_data_service__WEBPACK_IMPORTED_MODULE_3__["DataService"], _angular_router__WEBPACK_IMPORTED_MODULE_2__["ActivatedRoute"]])
@@ -381,17 +573,6 @@ var BlogItemDetailsComponent = /** @class */ (function () {
 /***/ (function(module, exports) {
 
 module.exports = "img{\r\n    width: 40rem;\r\n    \r\n}\r\n/*# sourceMappingURL=data:application/json;base64,eyJ2ZXJzaW9uIjozLCJzb3VyY2VzIjpbInNyYy9hcHAvY29tcG9uZW50cy9ibG9nLWl0ZW0taW1hZ2UvYmxvZy1pdGVtLWltYWdlLmNvbXBvbmVudC5jc3MiXSwibmFtZXMiOltdLCJtYXBwaW5ncyI6IkFBQUE7SUFDSSxZQUFZOztBQUVoQiIsImZpbGUiOiJzcmMvYXBwL2NvbXBvbmVudHMvYmxvZy1pdGVtLWltYWdlL2Jsb2ctaXRlbS1pbWFnZS5jb21wb25lbnQuY3NzIiwic291cmNlc0NvbnRlbnQiOlsiaW1ne1xyXG4gICAgd2lkdGg6IDQwcmVtO1xyXG4gICAgXHJcbn0iXX0= */"
-
-/***/ }),
-
-/***/ "./src/app/components/blog-item-image/blog-item-image.component.html":
-/*!***************************************************************************!*\
-  !*** ./src/app/components/blog-item-image/blog-item-image.component.html ***!
-  \***************************************************************************/
-/*! no static exports found */
-/***/ (function(module, exports) {
-
-module.exports = "<div class=\"card-img-top\">\r\n<img [src]=\"image\">\r\n</div>"
 
 /***/ }),
 
@@ -421,7 +602,7 @@ var BlogItemImageComponent = /** @class */ (function () {
     BlogItemImageComponent = tslib__WEBPACK_IMPORTED_MODULE_0__["__decorate"]([
         Object(_angular_core__WEBPACK_IMPORTED_MODULE_1__["Component"])({
             selector: 'blog-item-image',
-            template: __webpack_require__(/*! ./blog-item-image.component.html */ "./src/app/components/blog-item-image/blog-item-image.component.html"),
+            template: __webpack_require__(/*! raw-loader!./blog-item-image.component.html */ "./node_modules/raw-loader/index.js!./src/app/components/blog-item-image/blog-item-image.component.html"),
             styles: [__webpack_require__(/*! ./blog-item-image.component.css */ "./src/app/components/blog-item-image/blog-item-image.component.css")]
         }),
         tslib__WEBPACK_IMPORTED_MODULE_0__["__metadata"]("design:paramtypes", [])
@@ -441,17 +622,6 @@ var BlogItemImageComponent = /** @class */ (function () {
 /***/ (function(module, exports) {
 
 module.exports = "\n/*# sourceMappingURL=data:application/json;base64,eyJ2ZXJzaW9uIjozLCJzb3VyY2VzIjpbXSwibmFtZXMiOltdLCJtYXBwaW5ncyI6IiIsImZpbGUiOiJzcmMvYXBwL2NvbXBvbmVudHMvYmxvZy1pdGVtLXRleHQvYmxvZy1pdGVtLXRleHQuY29tcG9uZW50LmNzcyJ9 */"
-
-/***/ }),
-
-/***/ "./src/app/components/blog-item-text/blog-item-text.component.html":
-/*!*************************************************************************!*\
-  !*** ./src/app/components/blog-item-text/blog-item-text.component.html ***!
-  \*************************************************************************/
-/*! no static exports found */
-/***/ (function(module, exports) {
-
-module.exports = "<p class=\"card text-center\">\r\n  {{ text | summary:35 }}\r\n</p>\r\n\r\n<a [routerLink]=\"['/blog/detail/',id]\" class=\"btn btn-primary\">\r\n  Więcej\r\n </a>\r\n "
 
 /***/ }),
 
@@ -485,7 +655,7 @@ var BlogItemTextComponent = /** @class */ (function () {
     BlogItemTextComponent = tslib__WEBPACK_IMPORTED_MODULE_0__["__decorate"]([
         Object(_angular_core__WEBPACK_IMPORTED_MODULE_1__["Component"])({
             selector: 'blog-item-text',
-            template: __webpack_require__(/*! ./blog-item-text.component.html */ "./src/app/components/blog-item-text/blog-item-text.component.html"),
+            template: __webpack_require__(/*! raw-loader!./blog-item-text.component.html */ "./node_modules/raw-loader/index.js!./src/app/components/blog-item-text/blog-item-text.component.html"),
             styles: [__webpack_require__(/*! ./blog-item-text.component.css */ "./src/app/components/blog-item-text/blog-item-text.component.css")]
         }),
         tslib__WEBPACK_IMPORTED_MODULE_0__["__metadata"]("design:paramtypes", [])
@@ -505,17 +675,6 @@ var BlogItemTextComponent = /** @class */ (function () {
 /***/ (function(module, exports) {
 
 module.exports = "\n/*# sourceMappingURL=data:application/json;base64,eyJ2ZXJzaW9uIjozLCJzb3VyY2VzIjpbXSwibmFtZXMiOltdLCJtYXBwaW5ncyI6IiIsImZpbGUiOiJzcmMvYXBwL2NvbXBvbmVudHMvYmxvZy1pdGVtL2Jsb2ctaXRlbS5jb21wb25lbnQuY3NzIn0= */"
-
-/***/ }),
-
-/***/ "./src/app/components/blog-item/blog-item.component.html":
-/*!***************************************************************!*\
-  !*** ./src/app/components/blog-item/blog-item.component.html ***!
-  \***************************************************************/
-/*! no static exports found */
-/***/ (function(module, exports) {
-
-module.exports = "<blog-item-image [image]=\"image\"></blog-item-image>\r\n<blog-item-text [id]=\"id\" [text]=\"text\"></blog-item-text>"
 
 /***/ }),
 
@@ -553,7 +712,7 @@ var BlogItemComponent = /** @class */ (function () {
     BlogItemComponent = tslib__WEBPACK_IMPORTED_MODULE_0__["__decorate"]([
         Object(_angular_core__WEBPACK_IMPORTED_MODULE_1__["Component"])({
             selector: 'blog-item',
-            template: __webpack_require__(/*! ./blog-item.component.html */ "./src/app/components/blog-item/blog-item.component.html"),
+            template: __webpack_require__(/*! raw-loader!./blog-item.component.html */ "./node_modules/raw-loader/index.js!./src/app/components/blog-item/blog-item.component.html"),
             styles: [__webpack_require__(/*! ./blog-item.component.css */ "./src/app/components/blog-item/blog-item.component.css")]
         }),
         tslib__WEBPACK_IMPORTED_MODULE_0__["__metadata"]("design:paramtypes", [])
@@ -573,17 +732,6 @@ var BlogItemComponent = /** @class */ (function () {
 /***/ (function(module, exports) {
 
 module.exports = "\n/*# sourceMappingURL=data:application/json;base64,eyJ2ZXJzaW9uIjozLCJzb3VyY2VzIjpbXSwibmFtZXMiOltdLCJtYXBwaW5ncyI6IiIsImZpbGUiOiJzcmMvYXBwL2NvbXBvbmVudHMvYmxvZy9ibG9nLmNvbXBvbmVudC5jc3MifQ== */"
-
-/***/ }),
-
-/***/ "./src/app/components/blog/blog.component.html":
-/*!*****************************************************!*\
-  !*** ./src/app/components/blog/blog.component.html ***!
-  \*****************************************************/
-/*! no static exports found */
-/***/ (function(module, exports) {
-
-module.exports = "<selectize></selectize>\r\n<div *ngFor=\"let item of items$ | filter: filterText\" class=\"card-deck\">\r\n  <blog-item [id]=\"item.id\" [image]=\"item.image\" [text]=\"item.title\">\r\n    \r\n  </blog-item>\r\n</div>"
 
 /***/ }),
 
@@ -631,7 +779,7 @@ var BlogComponent = /** @class */ (function () {
     BlogComponent = tslib__WEBPACK_IMPORTED_MODULE_0__["__decorate"]([
         Object(_angular_core__WEBPACK_IMPORTED_MODULE_1__["Component"])({
             selector: 'blog',
-            template: __webpack_require__(/*! ./blog.component.html */ "./src/app/components/blog/blog.component.html"),
+            template: __webpack_require__(/*! raw-loader!./blog.component.html */ "./node_modules/raw-loader/index.js!./src/app/components/blog/blog.component.html"),
             styles: [__webpack_require__(/*! ./blog.component.css */ "./src/app/components/blog/blog.component.css")]
         }),
         tslib__WEBPACK_IMPORTED_MODULE_0__["__metadata"]("design:paramtypes", [_services_data_service__WEBPACK_IMPORTED_MODULE_2__["DataService"], _angular_router__WEBPACK_IMPORTED_MODULE_3__["Router"]])
@@ -651,17 +799,6 @@ var BlogComponent = /** @class */ (function () {
 /***/ (function(module, exports) {
 
 module.exports = "\n/*# sourceMappingURL=data:application/json;base64,eyJ2ZXJzaW9uIjozLCJzb3VyY2VzIjpbXSwibmFtZXMiOltdLCJtYXBwaW5ncyI6IiIsImZpbGUiOiJzcmMvYXBwL2NvbXBvbmVudHMvY29udGFjdC9jb250YWN0LmNvbXBvbmVudC5jc3MifQ== */"
-
-/***/ }),
-
-/***/ "./src/app/components/contact/contact.component.html":
-/*!***********************************************************!*\
-  !*** ./src/app/components/contact/contact.component.html ***!
-  \***********************************************************/
-/*! no static exports found */
-/***/ (function(module, exports) {
-
-module.exports = "  <div class=\"navbar\" id=\"navbar\"></div>\r\n\r\n  <form>\r\n      <div class=\"container\">\r\n          <div class=\"form-group\">\r\n              <label for=\"firstName\">First name</label>\r\n              <input type=\"text\" class=\"form-control\" id=\"exampleInputEmail1\" aria-describedby=\"emailHelp\"\r\n                  placeholder=\"Enter first name\">\r\n          </div>\r\n\r\n          <div class=\"form-group\">\r\n              <label for=\"firstName\">Last name</label>\r\n              <input type=\"text\" class=\"form-control\" id=\"exampleInputEmail1\" aria-describedby=\"emailHelp\"\r\n                  placeholder=\"Enter last name\">\r\n          </div>\r\n\r\n          <div class=\"form-group\">\r\n              <label for=\"exampleInputEmail1\">Email address</label>\r\n              <input type=\"email\" class=\"form-control\" id=\"exampleInputEmail1\" aria-describedby=\"emailHelp\"\r\n                  placeholder=\"Enter email\">\r\n              <small id=\"emailHelp\" class=\"form-text text-muted\">We'll never share your email with anyone\r\n                  else.</small>\r\n          </div>\r\n\r\n\r\n          <div class=\"form-check form-check-inline\">\r\n              <input class=\"form-check-input\" type=\"radio\" name=\"exampleRadios\" id=\"man\" value=\"option1\" checked>\r\n              <label class=\"form-check-label\" for=\"man\">\r\n                  Man\r\n              </label>\r\n          </div>\r\n\r\n          <div class=\"form-check form-check-inline\">\r\n              <input class=\"form-check-input\" type=\"radio\" name=\"exampleRadios\" id=\"woman\" value=\"option2\">\r\n              <label class=\"form-check-label\" for=\"woman\">\r\n                  Woman\r\n              </label>\r\n          </div>\r\n\r\n          <div class=\"form-group\">\r\n              <label for=\"exampleFormControlTextarea1\">Opis</label>\r\n              <textarea class=\"form-control\" id=\"exampleFormControlTextarea1\" rows=\"3\"></textarea>\r\n          </div>\r\n\r\n          <div class=\"form-group\">\r\n              <label for=\"customRange1\">Age</label>\r\n              <select class=\"custom-select\">\r\n                  <option selected>choose age</option>\r\n                  <option value=\"1\">>18</option>\r\n                  <option value=\"2\">18-25</option>\r\n                  <option value=\"3\">25-50</option>\r\n                  <option value=\"4\"><50</option> \r\n              </select> \r\n          </div> \r\n\r\n          <div class=\"form-group\">\r\n              <label for=\"inputState\">Marital status</label>\r\n              <select id=\"inputState\" class=\"form-control\">\r\n                  <option selected>single</option>\r\n                  <option>married</option>\r\n                  <option>divorced</option>\r\n                  <option>widow</option>\r\n              </select>\r\n          </div>\r\n\r\n          <div>\r\n              <button type=\"submit\" class=\"btn btn-primary\">Submit</button>\r\n          \r\n              <button type=\"reset\" class=\"btn btn-primary\">Reset</button>\r\n          </div>\r\n      </div>\r\n\r\n      <div class=\"scroll\">\r\n          <i class=\"fa fa-arrow-circle-up\" aria-hidden=\"true\">\r\n          <button class=\"btn btn-success\">Do góry</button>\r\n          </i>\r\n      </div>\r\n       \r\n  </form>\r\n"
 
 /***/ }),
 
@@ -687,7 +824,7 @@ var ContactComponent = /** @class */ (function () {
     ContactComponent = tslib__WEBPACK_IMPORTED_MODULE_0__["__decorate"]([
         Object(_angular_core__WEBPACK_IMPORTED_MODULE_1__["Component"])({
             selector: 'contact',
-            template: __webpack_require__(/*! ./contact.component.html */ "./src/app/components/contact/contact.component.html"),
+            template: __webpack_require__(/*! raw-loader!./contact.component.html */ "./node_modules/raw-loader/index.js!./src/app/components/contact/contact.component.html"),
             styles: [__webpack_require__(/*! ./contact.component.css */ "./src/app/components/contact/contact.component.css")]
         }),
         tslib__WEBPACK_IMPORTED_MODULE_0__["__metadata"]("design:paramtypes", [])
@@ -707,17 +844,6 @@ var ContactComponent = /** @class */ (function () {
 /***/ (function(module, exports) {
 
 module.exports = "\n/*# sourceMappingURL=data:application/json;base64,eyJ2ZXJzaW9uIjozLCJzb3VyY2VzIjpbXSwibmFtZXMiOltdLCJtYXBwaW5ncyI6IiIsImZpbGUiOiJzcmMvYXBwL2NvbXBvbmVudHMvaG9tZS9ob21lLmNvbXBvbmVudC5jc3MifQ== */"
-
-/***/ }),
-
-/***/ "./src/app/components/home/home.component.html":
-/*!*****************************************************!*\
-  !*** ./src/app/components/home/home.component.html ***!
-  \*****************************************************/
-/*! no static exports found */
-/***/ (function(module, exports) {
-
-module.exports = "  <div class=\"navbar\" id=\"navbar\"></div>\r\n    \r\n    <div class=\"row\">\r\n        <div class=\"col-9\">Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Hendrerit dolor magna eget est lorem ipsum. Duis at consectetur lorem donec massa. Feugiat in fermentum posuere urna nec. Elementum tempus egestas sed sed risus pretium quam vulputate dignissim. Orci porta non pulvinar neque laoreet suspendisse interdum consectetur. Ligula ullamcorper malesuada proin libero nunc. Non sodales neque sodales ut etiam sit. Scelerisque eleifend donec pretium vulputate sapien nec. Felis donec et odio pellentesque diam volutpat commodo sed. Enim blandit volutpat maecenas volutpat blandit aliquam etiam erat velit.\r\n\r\n                Consequat mauris nunc congue nisi vitae suscipit tellus. Natoque penatibus et magnis dis parturient. Bibendum ut tristique et egestas quis ipsum suspendisse ultrices. Maecenas volutpat blandit aliquam etiam. Varius vel pharetra vel turpis nunc eget lorem dolor sed. Volutpat est velit egestas dui id ornare arcu odio ut. Ut sem viverra aliquet eget sit amet tellus cras. Eget felis eget nunc lobortis. Faucibus in ornare quam viverra orci sagittis eu. Aliquam purus sit amet luctus venenatis lectus magna fringilla. Netus et malesuada fames ac turpis egestas. Non arcu risus quis varius quam quisque.\r\n                \r\n                Vitae proin sagittis nisl rhoncus mattis. Eget est lorem ipsum dolor sit amet consectetur adipiscing elit. Lacus sed turpis tincidunt id aliquet risus. Penatibus et magnis dis parturient montes nascetur ridiculus mus mauris. Malesuada fames ac turpis egestas integer eget. Semper quis lectus nulla at volutpat diam ut. Purus non enim praesent elementum facilisis. Porttitor lacus luctus accumsan tortor posuere ac. Accumsan in nisl nisi scelerisque eu. Nibh tellus molestie nunc non. Ut morbi tincidunt augue interdum velit euismod. Tellus integer feugiat scelerisque varius morbi enim nunc. Id aliquet lectus proin nibh nisl condimentum. Lacinia quis vel eros donec ac. Purus in massa tempor nec feugiat. Tellus rutrum tellus pellentesque eu. Pretium viverra suspendisse potenti nullam ac tortor vitae purus. Donec ultrices tincidunt arcu non sodales neque sodales ut. Nisl tincidunt eget nullam non. Sit amet consectetur adipiscing elit ut aliquam purus sit.\r\n                \r\n                Nisi scelerisque eu ultrices vitae auctor eu augue ut. Odio eu feugiat pretium nibh. Amet consectetur adipiscing elit ut aliquam purus. Ullamcorper malesuada proin libero nunc consequat interdum varius sit. Egestas congue quisque egestas diam in arcu cursus. Ut diam quam nulla porttitor massa id neque aliquam vestibulum. Scelerisque eu ultrices vitae auctor eu augue ut lectus. Sed enim ut sem viverra aliquet eget sit amet tellus. Sit amet consectetur adipiscing elit pellentesque habitant morbi tristique senectus. Massa enim nec dui nunc mattis. Mattis aliquam faucibus purus in massa tempor nec feugiat nisl. Malesuada fames ac turpis egestas maecenas. Tortor posuere ac ut consequat semper viverra. Praesent tristique magna sit amet purus gravida. In fermentum et sollicitudin ac orci. Pretium fusce id velit ut tortor pretium viverra suspendisse potenti. Quis lectus nulla at volutpat diam ut venenatis tellus. Tempor orci eu lobortis elementum nibh tellus molestie. Nunc eget lorem dolor sed viverra.\r\n                \r\n                Facilisi cras fermentum odio eu feugiat. Ornare arcu dui vivamus arcu felis bibendum. Curabitur gravida arcu ac tortor dignissim convallis aenean. Venenatis tellus in metus vulputate eu. Eros donec ac odio tempor orci dapibus ultrices in. Nunc sed blandit libero volutpat sed cras. Mauris rhoncus aenean vel elit scelerisque mauris pellentesque pulvinar. Id diam vel quam elementum pulvinar. At lectus urna duis convallis convallis tellus id interdum velit. Rutrum tellus pellentesque eu tincidunt tortor aliquam nulla facilisi. Adipiscing elit duis tristique sollicitudin nibh sit amet commodo. Pharetra pharetra massa massa ultricies mi quis. Imperdiet proin fermentum leo vel. Justo nec ultrices dui sapien eget mi proin. Convallis a cras semper auctor neque vitae tempus. Morbi tristique senectus et netus et. Nulla facilisi nullam vehicula ipsum a arcu. Nisl nisi scelerisque eu ultrices vitae auctor eu.\r\n                \r\n                Curabitur gravida arcu ac tortor. Pulvinar proin gravida hendrerit lectus. Ut faucibus pulvinar elementum integer enim neque volutpat ac. Ornare arcu dui vivamus arcu felis bibendum ut. Leo vel fringilla est ullamcorper eget nulla facilisi. Lectus magna fringilla urna porttitor rhoncus dolor. Venenatis lectus magna fringilla urna porttitor rhoncus. Eu sem integer vitae justo eget. Vulputate sapien nec sagittis aliquam malesuada bibendum. Arcu ac tortor dignissim convallis aenean. Morbi tincidunt augue interdum velit euismod. Condimentum id venenatis a condimentum vitae sapien pellentesque. Ultrices eros in cursus turpis massa tincidunt dui ut ornare. Ac turpis egestas sed tempus urna et pharetra pharetra massa. Sit amet consectetur adipiscing elit ut. Ultrices neque ornare aenean euismod. Eget lorem dolor sed viverra ipsum nunc aliquet.\r\n                \r\n                Scelerisque eu ultrices vitae auctor eu augue ut lectus. Non quam lacus suspendisse faucibus. Diam sollicitudin tempor id eu nisl nunc. Nulla aliquet porttitor lacus luctus. Vulputate mi sit amet mauris. Egestas integer eget aliquet nibh. Vestibulum lectus mauris ultrices eros in cursus turpis massa. Nibh tellus molestie nunc non blandit massa enim nec. Et molestie ac feugiat sed lectus vestibulum mattis ullamcorper. Id cursus metus aliquam eleifend mi in nulla posuere. Integer eget aliquet nibh praesent.\r\n                \r\n                Pellentesque eu tincidunt tortor aliquam nulla. Sed sed risus pretium quam vulputate dignissim suspendisse in est. Sem fringilla ut morbi tincidunt augue interdum velit. Ultrices in iaculis nunc sed augue lacus viverra vitae congue. Netus et malesuada fames ac turpis egestas. Hendrerit gravida rutrum quisque non tellus orci ac. Amet luctus venenatis lectus magna fringilla urna. Vitae sapien pellentesque habitant morbi tristique senectus et. Nullam non nisi est sit amet facilisis. Feugiat nibh sed pulvinar proin. Id cursus metus aliquam eleifend mi in. Pulvinar pellentesque habitant morbi tristique. Fringilla est ullamcorper eget nulla facilisi etiam dignissim diam quis. Cursus eget nunc scelerisque viverra. Eu volutpat odio facilisis mauris sit. Dapibus ultrices in iaculis nunc. Neque ornare aenean euismod elementum. Justo eget magna fermentum iaculis eu. Quam pellentesque nec nam aliquam sem et tortor. Magna fringilla urna porttitor rhoncus.\r\n                \r\n                Pulvinar mattis nunc sed blandit libero volutpat. Laoreet sit amet cursus sit amet dictum sit. Pharetra magna ac placerat vestibulum lectus mauris ultrices eros. Placerat duis ultricies lacus sed turpis. Volutpat maecenas volutpat blandit aliquam. Aliquam faucibus purus in massa tempor nec feugiat nisl pretium. Morbi tristique senectus et netus et malesuada fames ac turpis. Risus nec feugiat in fermentum posuere. Nec tincidunt praesent semper feugiat nibh sed pulvinar proin. Nibh nisl condimentum id venenatis. Pharetra sit amet aliquam id diam. Elementum curabitur vitae nunc sed velit dignissim. Diam maecenas ultricies mi eget. Tortor vitae purus faucibus ornare suspendisse sed nisi lacus sed. Egestas quis ipsum suspendisse ultrices gravida dictum fusce. Mauris ultrices eros in cursus turpis massa tincidunt. Accumsan tortor posuere ac ut. Fames ac turpis egestas integer eget aliquet nibh praesent tristique.\r\n                \r\n                Elementum nisi quis eleifend quam adipiscing vitae. Pulvinar neque laoreet suspendisse interdum. Cursus vitae congue mauris rhoncus. Molestie a iaculis at erat pellentesque. Quis enim lobortis scelerisque fermentum dui faucibus in ornare quam. Nec nam aliquam sem et tortor. Ullamcorper eget nulla facilisi etiam dignissim diam quis enim lobortis. Dui nunc mattis enim ut. Leo a diam sollicitudin tempor. Enim nunc faucibus a pellentesque sit amet porttitor eget dolor. Senectus et netus et malesuada fames ac turpis egestas maecenas. At lectus urna duis convallis convallis tellus id. Neque vitae tempus quam pellentesque nec. Nulla porttitor massa id neque aliquam vestibulum morbi. Mauris nunc congue nisi vitae suscipit tellus mauris a diam. Mauris cursus mattis molestie a iaculis.\r\n                \r\n                Justo eget magna fermentum iaculis eu non. Vulputate mi sit amet mauris. Vulputate dignissim suspendisse in est ante in nibh. Justo donec enim diam vulputate ut pharetra sit amet. Elementum nisi quis eleifend quam. Ipsum faucibus vitae aliquet nec. Sit amet consectetur adipiscing elit duis tristique. Id consectetur purus ut faucibus pulvinar elementum integer enim. Nunc sed velit dignissim sodales ut eu. Cras pulvinar mattis nunc sed blandit libero volutpat sed cras. Lacus luctus accumsan tortor posuere ac ut consequat semper viverra. Pellentesque habitant morbi tristique senectus. Amet nulla facilisi morbi tempus. Porta nibh venenatis cras sed. Lectus quam id leo in vitae turpis. Vivamus arcu felis bibendum ut tristique et egestas quis. Accumsan lacus vel facilisis volutpat est velit. Erat nam at lectus urna duis convallis convallis.\r\n                \r\n                Consequat mauris nunc congue nisi vitae suscipit tellus. In ante metus dictum at tempor commodo ullamcorper. Eget arcu dictum varius duis at consectetur. At urna condimentum mattis pellentesque id nibh tortor. Purus gravida quis blandit turpis cursus in. Dui vivamus arcu felis bibendum ut tristique et egestas. Nulla facilisi cras fermentum odio eu feugiat. Magna ac placerat vestibulum lectus mauris ultrices. Tellus pellentesque eu tincidunt tortor aliquam nulla facilisi cras fermentum. Iaculis urna id volutpat lacus laoreet non curabitur gravida. Aliquet risus feugiat in ante metus dictum at tempor. Diam volutpat commodo sed egestas. Eget lorem dolor sed viverra ipsum nunc aliquet bibendum enim. Nulla facilisi etiam dignissim diam quis enim. Arcu vitae elementum curabitur vitae. Hendrerit gravida rutrum quisque non tellus orci ac auctor. Ornare arcu odio ut sem nulla pharetra. Sollicitudin aliquam ultrices sagittis orci a scelerisque purus semper eget. Et leo duis ut diam quam nulla.\r\n                \r\n                Non consectetur a erat nam at lectus urna duis. Gravida neque convallis a cras semper auctor neque vitae. Sed velit dignissim sodales ut eu sem integer vitae justo. Facilisis leo vel fringilla est ullamcorper. Semper quis lectus nulla at volutpat diam ut venenatis. Scelerisque eleifend donec pretium vulputate sapien nec. Mauris pharetra et ultrices neque ornare aenean euismod elementum. Est placerat in egestas erat imperdiet sed euismod nisi. Nunc vel risus commodo viverra maecenas accumsan lacus vel facilisis. Amet consectetur adipiscing elit duis tristique sollicitudin nibh. Maecenas sed enim ut sem. Sagittis eu volutpat odio facilisis mauris.\r\n                \r\n                Egestas congue quisque egestas diam in arcu cursus. Sit amet purus gravida quis blandit turpis cursus in. Egestas egestas fringilla phasellus faucibus scelerisque eleifend donec pretium vulputate. Lobortis scelerisque fermentum dui faucibus in ornare quam. Turpis egestas pretium aenean pharetra magna. Amet facilisis magna etiam tempor orci eu lobortis elementum nibh. Amet est placerat in egestas erat imperdiet sed euismod. Sed vulputate mi sit amet mauris commodo. Sagittis aliquam malesuada bibendum arcu vitae elementum. Nunc sed blandit libero volutpat sed. Fringilla est ullamcorper eget nulla facilisi etiam. Mi tempus imperdiet nulla malesuada pellentesque. Lobortis elementum nibh tellus molestie nunc. Feugiat nisl pretium fusce id velit ut. Vel pharetra vel turpis nunc eget lorem. Arcu non sodales neque sodales ut etiam.\r\n                \r\n                Mauris augue neque gravida in fermentum et. Porttitor rhoncus dolor purus non enim. Consequat mauris nunc congue nisi vitae suscipit. Eget gravida cum sociis natoque. Phasellus vestibulum lorem sed risus ultricies tristique nulla. Velit dignissim sodales ut eu sem integer vitae justo. Aliquet sagittis id consectetur purus ut faucibus. Donec massa sapien faucibus et molestie. Quam lacus suspendisse faucibus interdum posuere lorem. Semper auctor neque vitae tempus quam pellentesque nec nam aliquam. Fermentum dui faucibus in ornare quam viverra orci sagittis. A lacus vestibulum sed arcu non odio euismod.\r\n                \r\n                Aliquam ut porttitor leo a diam. Leo integer malesuada nunc vel risus commodo. Tempor orci eu lobortis elementum nibh. At risus viverra adipiscing at in tellus integer feugiat scelerisque. Nibh cras pulvinar mattis nunc sed blandit libero volutpat. Viverra accumsan in nisl nisi scelerisque eu ultrices vitae auctor. Adipiscing enim eu turpis egestas pretium aenean pharetra magna ac. In dictum non consectetur a erat nam at. Integer malesuada nunc vel risus commodo viverra maecenas. Platea dictumst quisque sagittis purus sit. Aliquet enim tortor at auctor urna nunc id cursus. Scelerisque varius morbi enim nunc faucibus a. Sapien nec sagittis aliquam malesuada bibendum arcu vitae elementum curabitur. Urna condimentum mattis pellentesque id nibh tortor. Justo laoreet sit amet cursus sit amet dictum. Ridiculus mus mauris vitae ultricies leo integer. Ac ut consequat semper viverra nam libero. Vel facilisis volutpat est velit egestas dui. Vehicula ipsum a arcu cursus. Nunc congue nisi vitae suscipit tellus mauris a.\r\n                \r\n                Consequat id porta nibh venenatis cras sed felis. Eu volutpat odio facilisis mauris sit amet. A erat nam at lectus urna duis convallis convallis tellus. Pretium viverra suspendisse potenti nullam. Sed ullamcorper morbi tincidunt ornare massa eget egestas purus. Ut tristique et egestas quis. Neque viverra justo nec ultrices dui sapien eget mi. Aenean et tortor at risus. Et leo duis ut diam quam nulla. Suspendisse faucibus interdum posuere lorem ipsum dolor sit amet. Enim lobortis scelerisque fermentum dui faucibus in ornare quam. Parturient montes nascetur ridiculus mus mauris vitae ultricies leo. Sagittis aliquam malesuada bibendum arcu vitae. A diam maecenas sed enim. Aenean et tortor at risus viverra. Pulvinar pellentesque habitant morbi tristique senectus et netus et.\r\n                \r\n                Felis eget velit aliquet sagittis id. Sed vulputate mi sit amet mauris commodo quis imperdiet massa. Dui ut ornare lectus sit amet est placerat in. Convallis a cras semper auctor neque vitae tempus quam. Risus at ultrices mi tempus. Egestas pretium aenean pharetra magna. Massa enim nec dui nunc. In nibh mauris cursus mattis molestie a iaculis at erat. Elementum nisi quis eleifend quam adipiscing. Eget arcu dictum varius duis. Scelerisque varius morbi enim nunc faucibus a pellentesque sit. Lorem sed risus ultricies tristique nulla aliquet enim. Velit ut tortor pretium viverra. Vitae semper quis lectus nulla at volutpat diam ut. Erat pellentesque adipiscing commodo elit at imperdiet dui accumsan. Eget duis at tellus at urna.\r\n                \r\n                Accumsan tortor posuere ac ut consequat. Lacus vestibulum sed arcu non odio euismod lacinia at. Diam quam nulla porttitor massa id neque aliquam. Ac tortor vitae purus faucibus ornare suspendisse sed nisi lacus. Sed turpis tincidunt id aliquet risus feugiat. Mattis molestie a iaculis at erat pellentesque adipiscing commodo. Non sodales neque sodales ut etiam sit amet. Amet mattis vulputate enim nulla aliquet porttitor. Tristique risus nec feugiat in fermentum posuere urna. Elit scelerisque mauris pellentesque pulvinar pellentesque habitant morbi. Sit amet porttitor eget dolor morbi non.\r\n                \r\n                Enim ut tellus elementum sagittis vitae et leo duis. Eget lorem dolor sed viverra ipsum. Orci ac auctor augue mauris augue neque gravida in fermentum. Metus vulputate eu scelerisque felis imperdiet proin fermentum leo. Sed vulputate mi sit amet mauris. Vitae congue eu consequat ac felis donec. Odio ut sem nulla pharetra diam sit amet. Eget nulla facilisi etiam dignissim diam quis. Lorem mollis aliquam ut porttitor leo a diam sollicitudin tempor. In egestas erat imperdiet sed euismod nisi porta lorem mollis. Est lorem ipsum dolor sit amet consectetur adipiscing elit. Vitae congue mauris rhoncus aenean vel elit scelerisque. In pellentesque massa placerat duis ultricies lacus sed turpis tincidunt. Consequat id porta nibh venenatis cras. Amet risus nullam eget felis eget nunc lobortis. Ipsum faucibus vitae aliquet nec. Enim facilisis gravida neque convallis a. Dictumst quisque sagittis purus sit amet volutpat consequat mauris nunc.</div>\r\n        <div class=\"col-3 s-bg\">\r\n            <h3>Szybki kontakt</h3>\r\n            \r\n            <a href=\"#\"><i class=\"fab fa-facebook-square\"></i></a>\r\n            <a href=\"#\"><i class=\"fab fa-instagram\"></i></a>\r\n            <a href=\"#\"><i class=\"fas fa-at\"></i></a>\r\n        </div>\r\n    </div>\r\n\r\n    <div class=\"scroll\">\r\n            <button class=\"btn btn-success\" id=\"myButton\">\r\n                <i class=\"fa fa-arrow-circle-up\" aria-hidden=\"true\"></i>\r\n                 Do góry\r\n            </button>\r\n    </div>\r\n             \r\n    <div class=\"footer\">Copyrights © Sebastian Pulak | 2019</div>\r\n"
 
 /***/ }),
 
@@ -743,7 +869,7 @@ var HomeComponent = /** @class */ (function () {
     HomeComponent = tslib__WEBPACK_IMPORTED_MODULE_0__["__decorate"]([
         Object(_angular_core__WEBPACK_IMPORTED_MODULE_1__["Component"])({
             selector: 'home',
-            template: __webpack_require__(/*! ./home.component.html */ "./src/app/components/home/home.component.html"),
+            template: __webpack_require__(/*! raw-loader!./home.component.html */ "./node_modules/raw-loader/index.js!./src/app/components/home/home.component.html"),
             styles: [__webpack_require__(/*! ./home.component.css */ "./src/app/components/home/home.component.css")]
         }),
         tslib__WEBPACK_IMPORTED_MODULE_0__["__metadata"]("design:paramtypes", [])
@@ -763,17 +889,6 @@ var HomeComponent = /** @class */ (function () {
 /***/ (function(module, exports) {
 
 module.exports = "\n/*# sourceMappingURL=data:application/json;base64,eyJ2ZXJzaW9uIjozLCJzb3VyY2VzIjpbXSwibmFtZXMiOltdLCJtYXBwaW5ncyI6IiIsImZpbGUiOiJzcmMvYXBwL2NvbXBvbmVudHMvcXVpei9xdWl6LmNvbXBvbmVudC5jc3MifQ== */"
-
-/***/ }),
-
-/***/ "./src/app/components/quiz/quiz.component.html":
-/*!*****************************************************!*\
-  !*** ./src/app/components/quiz/quiz.component.html ***!
-  \*****************************************************/
-/*! no static exports found */
-/***/ (function(module, exports) {
-
-module.exports = "\r\n  <div class=\"navbar\" id=\"navbar\"></div>\r\n<div class=\"row\"></div>\r\n<div class=\"container list\">\r\n\r\n    Question: <span class=\"index\"></span>1<h4 class=\"question\"></h4>\r\n\r\n    Answers:\r\n    <ul class=\"list-group\">\r\n        <li class=\"list-group-item\"></li>\r\n        <li class=\"list-group-item\"></li>\r\n        <li class=\"list-group-item\"></li>\r\n        <li class=\"list-group-item\"></li>\r\n    </ul>\r\n\r\n    <div class=\"buttons\">\r\n        <button class=\"btn btn-danger previous\">Previous</button>\r\n        <button class=\"btn btn-success next\">Next</button>\r\n    </div>\r\n\r\n    <h4>Your score: <span class=\"score\">0</span>/20 points.</h4>\r\n</div>\r\n<div class=\"container results\" style=\"display: none\">\r\n    <div class=\"btn-group-custom\">\r\n        <button class=\"btn btn-success btn-custom restart\">Retake quiz</button>\r\n    </div>\r\n    <div class=\"score\">\r\n        <table class=\"table table-striped\">\r\n            <thead>\r\n            <tr>\r\n                <th>#</th>\r\n                <th>Point/20</th>\r\n                <th>Average</th>\r\n            </tr>\r\n            </thead>\r\n            <tbody>\r\n            <tr>\r\n                <th scope=\"row\">Your score</th>\r\n                <td class=\"userScorePoint\"></td>\r\n                <td class=\"average\"></td>\r\n            </tr>\r\n            </tbody>\r\n        </table>\r\n    </div>\r\n</div>\r\n\r\n\r\n"
 
 /***/ }),
 
@@ -799,7 +914,7 @@ var QuizComponent = /** @class */ (function () {
     QuizComponent = tslib__WEBPACK_IMPORTED_MODULE_0__["__decorate"]([
         Object(_angular_core__WEBPACK_IMPORTED_MODULE_1__["Component"])({
             selector: 'quiz',
-            template: __webpack_require__(/*! ./quiz.component.html */ "./src/app/components/quiz/quiz.component.html"),
+            template: __webpack_require__(/*! raw-loader!./quiz.component.html */ "./node_modules/raw-loader/index.js!./src/app/components/quiz/quiz.component.html"),
             styles: [__webpack_require__(/*! ./quiz.component.css */ "./src/app/components/quiz/quiz.component.css")]
         }),
         tslib__WEBPACK_IMPORTED_MODULE_0__["__metadata"]("design:paramtypes", [])
@@ -819,17 +934,6 @@ var QuizComponent = /** @class */ (function () {
 /***/ (function(module, exports) {
 
 module.exports = "\n/*# sourceMappingURL=data:application/json;base64,eyJ2ZXJzaW9uIjozLCJzb3VyY2VzIjpbXSwibmFtZXMiOltdLCJtYXBwaW5ncyI6IiIsImZpbGUiOiJzcmMvYXBwL2NvbXBvbmVudHMvc2VsZWN0aXplL3NlbGVjdGl6ZS5jb21wb25lbnQuY3NzIn0= */"
-
-/***/ }),
-
-/***/ "./src/app/components/selectize/selectize.component.html":
-/*!***************************************************************!*\
-  !*** ./src/app/components/selectize/selectize.component.html ***!
-  \***************************************************************/
-/*! no static exports found */
-/***/ (function(module, exports) {
-
-module.exports = "\r\n<input #input type=\"text\">\r\n\r\n<ul>\r\n  <li *ngFor=\"let p of posts$\">\r\n\t<a href=\"#\" [routerLink]=\"['/blog/detail/', p.id]\"> {{ p.text }}</a></li>\r\n</ul>\r\n"
 
 /***/ }),
 
@@ -865,13 +969,13 @@ var SelectizeComponent = /** @class */ (function () {
         });
     };
     tslib__WEBPACK_IMPORTED_MODULE_0__["__decorate"]([
-        Object(_angular_core__WEBPACK_IMPORTED_MODULE_1__["ViewChild"])('input'),
+        Object(_angular_core__WEBPACK_IMPORTED_MODULE_1__["ViewChild"])('input', { static: false }),
         tslib__WEBPACK_IMPORTED_MODULE_0__["__metadata"]("design:type", _angular_core__WEBPACK_IMPORTED_MODULE_1__["ElementRef"])
     ], SelectizeComponent.prototype, "input", void 0);
     SelectizeComponent = tslib__WEBPACK_IMPORTED_MODULE_0__["__decorate"]([
         Object(_angular_core__WEBPACK_IMPORTED_MODULE_1__["Component"])({
             selector: 'selectize',
-            template: __webpack_require__(/*! ./selectize.component.html */ "./src/app/components/selectize/selectize.component.html"),
+            template: __webpack_require__(/*! raw-loader!./selectize.component.html */ "./node_modules/raw-loader/index.js!./src/app/components/selectize/selectize.component.html"),
             styles: [__webpack_require__(/*! ./selectize.component.css */ "./src/app/components/selectize/selectize.component.css")]
         }),
         tslib__WEBPACK_IMPORTED_MODULE_0__["__metadata"]("design:paramtypes", [src_app_services_data_service__WEBPACK_IMPORTED_MODULE_4__["DataService"]])
@@ -891,17 +995,6 @@ var SelectizeComponent = /** @class */ (function () {
 /***/ (function(module, exports) {
 
 module.exports = "\n/*# sourceMappingURL=data:application/json;base64,eyJ2ZXJzaW9uIjozLCJzb3VyY2VzIjpbXSwibmFtZXMiOltdLCJtYXBwaW5ncyI6IiIsImZpbGUiOiJzcmMvYXBwL25hdmJhci9uYXZiYXIuY29tcG9uZW50LmNzcyJ9 */"
-
-/***/ }),
-
-/***/ "./src/app/navbar/navbar.component.html":
-/*!**********************************************!*\
-  !*** ./src/app/navbar/navbar.component.html ***!
-  \**********************************************/
-/*! no static exports found */
-/***/ (function(module, exports) {
-
-module.exports = "<nav>\r\n    <ul class=\"nav\">\r\n        <li class=\"nav-item\"><a class=\"nav-link active\" [routerLink]=\"['/']\">Home</a></li>\r\n        <li class=\"nav-item\"><a class=\"nav-link\" [routerLink]=\"['/blog']\">Blog</a></li>\r\n        <li class=\"nav-item\"><a class=\"nav-link\" [routerLink]=\"['/quiz']\">Quiz</a></li>\r\n        <li class=\"nav-item\"><a class=\"nav-link\" [routerLink]=\"['/contact']\">Contact</a></li>\r\n        <li class=\"nav-item\"><a class=\"nav-link\" [routerLink]=\"['/blog/detail/:id']\">Blog item details</a></li>\r\n    </ul>\r\n</nav>\r\n\r\n"
 
 /***/ }),
 
@@ -927,7 +1020,7 @@ var NavbarComponent = /** @class */ (function () {
     NavbarComponent = tslib__WEBPACK_IMPORTED_MODULE_0__["__decorate"]([
         Object(_angular_core__WEBPACK_IMPORTED_MODULE_1__["Component"])({
             selector: 'navbar',
-            template: __webpack_require__(/*! ./navbar.component.html */ "./src/app/navbar/navbar.component.html"),
+            template: __webpack_require__(/*! raw-loader!./navbar.component.html */ "./node_modules/raw-loader/index.js!./src/app/navbar/navbar.component.html"),
             styles: [__webpack_require__(/*! ./navbar.component.css */ "./src/app/navbar/navbar.component.css")]
         }),
         tslib__WEBPACK_IMPORTED_MODULE_0__["__metadata"]("design:paramtypes", [])
@@ -1045,6 +1138,9 @@ var DataService = /** @class */ (function () {
     DataService.prototype.getByText = function (data) {
         return this.http.get(this.url + '/api/posts/', data);
     };
+    DataService.prototype.createOrUpdate = function (post) {
+        return this.http.post(this.url + "/api/posts", post);
+    };
     DataService = tslib__WEBPACK_IMPORTED_MODULE_0__["__decorate"]([
         Object(_angular_core__WEBPACK_IMPORTED_MODULE_1__["Injectable"])(),
         tslib__WEBPACK_IMPORTED_MODULE_0__["__metadata"]("design:paramtypes", [_angular_common_http__WEBPACK_IMPORTED_MODULE_2__["HttpClient"]])
@@ -1064,17 +1160,6 @@ var DataService = /** @class */ (function () {
 /***/ (function(module, exports) {
 
 module.exports = "\n/*# sourceMappingURL=data:application/json;base64,eyJ2ZXJzaW9uIjozLCJzb3VyY2VzIjpbXSwibmFtZXMiOltdLCJtYXBwaW5ncyI6IiIsImZpbGUiOiJzcmMvYXBwL3NoYXJlZC9zZWFyY2gtYmFyL3NlYXJjaC1iYXIuY29tcG9uZW50LmNzcyJ9 */"
-
-/***/ }),
-
-/***/ "./src/app/shared/search-bar/search-bar.component.html":
-/*!*************************************************************!*\
-  !*** ./src/app/shared/search-bar/search-bar.component.html ***!
-  \*************************************************************/
-/*! no static exports found */
-/***/ (function(module, exports) {
-
-module.exports = "<input [(ngModel)]=\"filterText\"\r\n(ngModelChange)=\"sendFilter($event)\"\r\n      \tplaceholder=\"Wpisz nazwę...\" id=\"search\" type=\"search\"\r\nclass=\"form-control\">\r\n"
 
 /***/ }),
 
@@ -1118,7 +1203,7 @@ var SearchBarComponent = /** @class */ (function () {
     SearchBarComponent = tslib__WEBPACK_IMPORTED_MODULE_0__["__decorate"]([
         Object(_angular_core__WEBPACK_IMPORTED_MODULE_1__["Component"])({
             selector: 'search-bar',
-            template: __webpack_require__(/*! ./search-bar.component.html */ "./src/app/shared/search-bar/search-bar.component.html"),
+            template: __webpack_require__(/*! raw-loader!./search-bar.component.html */ "./node_modules/raw-loader/index.js!./src/app/shared/search-bar/search-bar.component.html"),
             styles: [__webpack_require__(/*! ./search-bar.component.css */ "./src/app/shared/search-bar/search-bar.component.css")]
         }),
         tslib__WEBPACK_IMPORTED_MODULE_0__["__metadata"]("design:paramtypes", [_angular_router__WEBPACK_IMPORTED_MODULE_2__["Router"], _angular_router__WEBPACK_IMPORTED_MODULE_2__["ActivatedRoute"]])
@@ -1191,7 +1276,7 @@ Object(_angular_platform_browser_dynamic__WEBPACK_IMPORTED_MODULE_1__["platformB
 /*! no static exports found */
 /***/ (function(module, exports, __webpack_require__) {
 
-module.exports = __webpack_require__(/*! D:\Pulak\TAI_LAB5\blog\src\main.ts */"./src/main.ts");
+module.exports = __webpack_require__(/*! F:\Uni\ang\TAI_LAB5\blog\src\main.ts */"./src/main.ts");
 
 
 /***/ })
