@@ -3,7 +3,7 @@ import {HttpClient} from '@angular/common/http';
 
 @Injectable()
 export class DataService {
-  private url = 'https://blogtai.herokuapp.com';
+  private url = 'http://localhost:3000';
 
 
   constructor(private http: HttpClient) {
